refactor: logger module was divided into separate parts
