+++ src/utils/index.js
@@ -1,0 +1,3 @@
+const getStringFromObject = require('./get-string-from-object');
+
+module.exports = { getStringFromObject };
